move repository to services dir
