+++ authark/application/domain/services/__init__.py
@@ -5,3 +5,4 @@ from .access_service import *
 from .verification_service import *
 from .enrollment_service import *
 from .identity_service import *
+from .repositories import *
