+++ authark/application/domain/services/access_service.py
@@ -1,6 +1,6 @@
 from typing import Dict, Any, List
 from ..models import User, Tenant, Token, Dominion
-from ..repositories import (
+from .repositories import (
     RankingRepository, RoleRepository, DominionRepository)
 from .token_service import AccessTokenService
