+++ authark/application/domain/services/repositories/__init__.py
@@ -1,4 +1,4 @@
-from ..common import QueryDomain
+from ...common import QueryDomain
 from .memory_model_repositories import (
     CredentialRepository, MemoryCredentialRepository,
     DominionRepository, MemoryDominionRepository,
