+++ authark/application/operation/managers/import_manager.py
@@ -1,6 +1,6 @@
 from typing import List, Optional, Any
 from ...domain.services import ImportService
-from ...domain.repositories import (
+from ...domain.services.repositories import (
     UserRepository, CredentialRepository, RoleRepository, RankingRepository,
     DominionRepository)
 from ...domain.models import User, Credential, Ranking, Dominion
